refactor: create solvers directory
