+++ src/solvers/monotonic_hill_climb.ts
